motor_car_v3/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
RPi.GPIO==0.6.3
Werkzeug==0.14.1
